Extracted parts of PrepareWorkspaceEvent into HostEvent
