+++ src/Event/PrepareWorkspaceEvent.php
@@ -2,40 +2,21 @@
 
 namespace Accompli\Event;
 
-use Accompli\Deployment\Host;
 use Accompli\Deployment\Workspace;
-use Symfony\Component\EventDispatcher\Event;
 
 /**
  * PrepareWorkspaceEvent.
  *
- * @author  Niels Nijens <nijens.niels@gmail.com>
+ * @author Niels Nijens <nijens.niels@gmail.com>
  */
-class PrepareWorkspaceEvent extends Event
+class PrepareWorkspaceEvent extends HostEvent
 {
-    /**
-     * The Host instance.
-     *
-     * @var Host
-     */
-    private $host;
-
     /**
      * The Workspace instance.
      *
      * @var Workspace
      */
     private $workspace;
-
-    /**
-     * Constructs a new PrepareWorkspaceEvent.
-     *
-     * @param Host $host
-     */
-    public function __construct(Host $host)
-    {
-        $this->host = $host;
-    }
 
     /**
      * Sets a Workspace instance.
@@ -45,16 +26,6 @@ class PrepareWorkspaceEvent extends Event
     public function setWorkspace(Workspace $workspace)
     {
         $this->workspace = $workspace;
-    }
-
-    /**
-     * Returns the Host instance.
-     *
-     * @return Host
-     */
-    public function getHost()
-    {
-        return $this->host;
     }
 
     /**
